Improve image compression
Previously, we'd jump from a 100% quality jpeg compressed bitmap to a 60% quality compressed bitmap at 1/4th the resolution. This meant that an image that was slightly over the auto-compression threshold would lose an unnecessarily large amount of resolution/quality.

This new method is more precise with how it compresses the images, and will bring the compressed size much closer to the target size