Fix crash on loading conversation
The switchMap was causing a crash because it was trying to use conversation as an observable, even though it is not. (caused by c3df5f231f625a77c2ef9bee5ea98fd7d598b038) There is no need for this code, because the whole function is now an observable.